+++ pkg/commands/apisecurity/tags/tags_test.go
@@ -12,7 +12,7 @@ import (
 	sub "github.com/fastly/cli/pkg/commands/apisecurity/tags"
 	fstfmt "github.com/fastly/cli/pkg/fmt"
 	"github.com/fastly/cli/pkg/testutil"
-	"github.com/fastly/go-fastly/v15/fastly/apisecurity/operations"
+	"github.com/fastly/go-fastly/v16/fastly/apisecurity/operations"
 )
 
 const (