+++ pkg/commands/authtoken/create.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/fastly/go-fastly/v15/fastly"
+	"github.com/fastly/go-fastly/v16/fastly"
 
 	"github.com/fastly/kingpin"