+++ pkg/commands/compute/compute_mocks_test.go
@@ -7,7 +7,7 @@ package compute_test
 import (
 	"context"
 
-	"github.com/fastly/go-fastly/v15/fastly"
+	"github.com/fastly/go-fastly/v16/fastly"
 
 	"github.com/fastly/cli/pkg/testutil"
 )