+++ pkg/commands/dashboard/printer/print.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"strings"
 
-	"github.com/fastly/go-fastly/v15/fastly"
+	"github.com/fastly/go-fastly/v16/fastly"
 
 	"github.com/fastly/cli/pkg/text"
 )