+++ pkg/commands/service/ratelimit/create.go
@@ -7,7 +7,7 @@ import (
 	"io"
 	"strings"
 
-	"github.com/fastly/go-fastly/v15/fastly"
+	"github.com/fastly/go-fastly/v16/fastly"
 
 	"4d63.com/optional"