+++ pkg/text/stats.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/fastly/go-fastly/v15/fastly"
+	"github.com/fastly/go-fastly/v16/fastly"
 )
 
 func PrintUsageTbl(out io.Writer, data *fastly.RegionsUsage) {